restructure namespace conforming to the namespace convention, which is a built-in convention in ASP.NET API Versioning.
