+++ RestApi.Test/V2/ExampleTest.cs
@@ -4,7 +4,7 @@ using System.Net;
 using System.Text.Json;
 using System.Threading.Tasks;
 using NUnit.Framework;
-using RestApi.Models.V2;
+using RestApi.Api.V2.Models;
 using RestApi.Test.Models;
 
 namespace RestApi.Test.V2
